refactor: move error types to dedicated module
